KVM: SEV: Allow for VMPL level specification in AP create
Update AP creation to support ADD/DESTROY of VMSAs at levels other than
VMPL0 in order to run with under an SVSM at VMPL1 or lower. To maintain
backwards compatibility, the VMPL is specified in bits 16 to 19 of the
AP Creation request in SW_EXITINFO1 of the GHCB.

In order to track the VMSAs at different levels, create arrays for the
VMSAs, GHCBs, registered GHCBs and others. When switching VMPL levels,
these entries will be used to set the VMSA and GHCB physical addresses
in the VMCB for the VMPL level.

In order ensure that the proper responses are returned in the proper GHCB,
the GHCB must be unmapped at the current level and saved for restoration
later when switching back to that VMPL level.

Additional checks are applied to prevent a non-VMPL0 vCPU from being able
to perform an AP creation request. Additionally, a vCPU cannot replace its
own VMSA.

Signed-off-by: Tom Lendacky <thomas.lendacky@amd.com>
[nstange@suse.de: forward-ported from svsm-preview-hv.
 sev_es_vcpu_reset(), formerly sev_es_vcpu_create(), became fragile
 now: the VMPL0 vmsa needs to get set before init_vmcb(), but the
 __svm_vcpu_reset() call cannot get moved to before the one to init_vmcb()
 in svm_vcpu_reset(). Should be working, but this deserves a proper cleanup.]
Signed-off-by: Nicolai Stange <nstange@suse.de>