x86/sev: add a SVSM vTPM platform device
If the SNP boot has a SVSM, probe for the vTPM device by sending a
SVSM_VTPM_QUERY call (function 8). The SVSM will return a bitmap with the
TPM_SEND_COMMAND bit set only if the vTPM is present and it is able to handle
TPM commands at runtime.

If a vTPM is found, register a platform device as "platform:tpm" so it
can be attached to the tpm_platform.c driver.

Signed-off-by: James Bottomley <James.Bottomley@HansenPartnership.com>
Signed-off-by: Joerg Roedel <jroedel@suse.de>
[SG] Code adjusted with some changes introduced in 6.11
Signed-off-by: Stefano Garzarella <sgarzare@redhat.com>