KVM: SEV: don't release previous VMSA page in __sev_snp_update_protected_guest_state()
The code in __sev_snp_update_protected_guest_state() supposed to release
the to be replaced VMSA is bogus: it inspects the replacement's backing memory
to decide about whether the previous, to be replaced VMSA should be freed
directly or not. More specifically, if the former is not backed by a restricted
memfd, it would free the latter directly. As the two VMSAs' backings are unrelated
this is wrong.

For AP creation, the code is dead anyway, because new VMSAs should always be
backed by private memory.

For AP destruction OTOH, the replacement, i.e. INVALID_PAGE, is certainly not backed
by some restricted memfd, and so the code would errorneously free up the previous,
to be replaced VMSA, irrespective of its backing. As the previous VMSA is still either
the original one installed at vCPU creation, or has subsequently been created by means
of an AP creation NAE and is backed by a restricted memfd, it must never get freed in
either case. Remove the corresponding logic in __sev_snp_update_protected_guest_state().

Signed-off-by: Nicolai Stange <nstange@suse.de>