KVM: TDX: Enumerate TD partitioning feature
TD partitioning is a feature introduced in TDX1.5, enables the creation
of multiple L2 VMs within a single TD. The maximum limit for L2 VMs is
set to three. This feature is exposed to the userspace VMM as part of
tdx_capabilities, allowing the userspace VMM to determine the number of
L2 VMs that a L1 TD will support. Each L2 VM requires an additional
number of pages for TDCS/TDVPS. These numbers are queried from the TDX
module and stored as part of tdx_info.

Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>