KVM: TDX: Manage the L2 SEPT paging pages for each L2 VM
The Host VMM is responsible for allocating paging pages for the L2 SEPT.
To facilitate the easy reclamation of these pages when a TD is
destroyed, a data structure called l2sept_header is introduced. This
structure records the Host Physical Address (HPA) of the L2 SEPT paging
pages. Each l2sept_header can be added to the l2sept_list corresponding
to its L2 VM.

Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>