KVM: TDX: Retry on the race of aug a page
When one vcpu is trying to send the seamcall TDH.MEM.PAGE.AUG to populate
SEPT, another may already be working in progress to create mapping for
the same gfn in SEPT as well. This may result in this vcpu receive the
error code TDX_OPERAND_BUSY, or TDX_EPT_ENTRY_STATE_INCORRECT with the
SEPT entry state already PRESENT. Just simply retry for these cases.

Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>