KVM: SVM: Add GUEST_TSC_FREQ MSR for Secure TSC enabled guests
Introduce the read-only MSR GUEST_TSC_FREQ (0xc0010134) that returns
guest's effective frequency in MHZ when Secure TSC is enabled for SNP
guests. Disable interception of this MSR when Secure TSC is enabled. Note
that GUEST_TSC_FREQ MSR is accessible only to the guest and not from the
hypervisor context.
(Vaishali: Adapt the value of MAX_DIRECT_ACCESS_MSRs to incorporate the
change introduced by this commit, allow Secure TSC to be used at VMPL 0
level)

Reviewed-by: Tom Lendacky <thomas.lendacky@amd.com>
Tested-by: Vaishali Thakkar <vaishali.thakkar@suse.com>
Signed-off-by: Nikunj A Dadhania <nikunj@amd.com>